% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\docType{data}
\name{meme}
\alias{meme}
\title{Initialize a MEME docker image}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
meme
}
\description{
Initializes a MEME docker image and calls specific methods
}
\keyword{datasets}
